Rename file_store file
