+++ linux/user-daemon/src/stores/mod.rs
@@ -1,6 +1,6 @@
 use u2f_core::{ApplicationKey, Counter};
 
-pub(crate) mod file;
+pub(crate) mod file_store;
 pub(crate) mod file_store_v2;
 pub(crate) mod secret_service;
